requirements.txt
# requirements.txt
setuptools
langchain
python-dotenv
wheel
spacy
transformers
streamlit
stqdm
pandas
spacy_streamlit